% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cellsize.R
\name{check_cellsize}
\alias{check_cellsize}
\title{check if the cellsize is appropriate}
\usage{
check_cellsize(object, cellsize)
}
\arguments{
\item{object}{the spatial object}

\item{cellsize}{the optional cellsize. If missing the function will return a relevant cellsize}
}
\description{
check if the cellsize is appropriate
}
