% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRTS.R, R/GRTSgrid.R, R/GRTSpoints.R,
%   R/GRTSpolygons.R, R/NdGRTS.R
\docType{methods}
\name{GRTS-methods}
\alias{GRTS-methods}
\alias{GRTS}
\alias{GRTS-methods}
\alias{GRTS,numeric-method}
\alias{GRTS,SpatialGrid-method}
\alias{GRTS,SpatialGridDataFrame-method}
\alias{GRTS,SpatialPoints-method}
\alias{GRTS,SpatialPointsDataFrame-method}
\alias{GRTS,SpatialPolygons-method}
\alias{GRTS,SpatialPolygonsDataFrame-method}
\alias{NdGRTS-methods}
\alias{NdGRTS}
\title{The GRTS wrapper function}
\usage{
GRTS(object, ...)

\S4method{GRTS}{numeric}(object, ...)

\S4method{GRTS}{SpatialGrid}(object, ...)

\S4method{GRTS}{SpatialGridDataFrame}(object, ...)

\S4method{GRTS}{SpatialPoints}(object, ...)

\S4method{GRTS}{SpatialPointsDataFrame}(object, ...)

\S4method{GRTS}{SpatialPolygons}(object, ...)

\S4method{GRTS}{SpatialPolygonsDataFrame}(object, ...)

NdGRTS(object, ...)
}
\arguments{
\item{object}{The object on which to generate a GRTS sample}

\item{...}{further arguments to the functions}

\item{object}{The object on which to generate the N dimensional GRTS ranking}

\item{...}{further arguments to the functions}
}
\value{
A GRTS sample

a data.frame with the design and ranking
}
\description{
The GRTS wrapper function selects the appropriate GRTS function based on the object.
}
\section{Methods}{

\describe{
 \item{\code{signature(object = "numeric")}}{Use \code{GRTS.default}}
 \item{\code{signature(object = "SpatialPoints")}}{Use \code{GRTS.point}}
 \item{\code{signature(object = "SpatialPointsDataFrame")}}{Use \code{GRTS.point}}
 \item{\code{signature(object = "SpatialGrid")}}{Use \code{GRTS.grid}}
 \item{\code{signature(object = "SpatialGridDataFrame")}}{Use \code{GRTS.grid}}
 \item{\code{signature(object = "SpatialPolygons")}}{Use \code{GRTS.polygon}}
 \item{\code{signature(object = "SpatialPolygonsDataFrame")}}{Use \code{GRTS.polygon}}
}
}

\examples{
 GRTS(4)
}
\seealso{
\code{\link{GRTS.default}}

\code{\link{GRTS.grid}}

\code{\link{GRTS.point}}

\code{\link{GRTS.polygon}}
}
\keyword{methods}
