% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_grts.R, R/nd_grts-list.R,
%   R/nd_grts-data_frame.R, R/nd_grts-spatialpoints.R,
%   R/nd_grts-spatialpointsdataframe.R
\docType{methods}
\name{nd_grts-methods}
\alias{nd_grts-methods}
\alias{nd_grts}
\alias{nd_grts-methods}
\alias{nd_grts,list-method}
\alias{nd_grts-methods}
\alias{nd_grts,data.frame-method}
\alias{nd_grts-methods}
\alias{nd_grts,SpatialPoints-method}
\alias{nd_grts-methods}
\alias{nd_grts,SpatialPointsDataFrame-method}
\title{Calculate the N dimensional GRTS ranking for the object}
\usage{
nd_grts(object, ...)

\S4method{nd_grts}{list}(object, ...)

\S4method{nd_grts}{data.frame}(object, ...)

\S4method{nd_grts}{SpatialPoints}(object, ...)

\S4method{nd_grts}{SpatialPointsDataFrame}(object, ...)
}
\arguments{
\item{object}{The object on which to generate the N dimensional GRTS ranking}

\item{...}{further arguments to the functions. See Details}
}
\value{
a data.frame with the design and ranking
}
\description{
While GRTS() works only on 2 dimensional data, nd_grts() can work with several dimensions, starting from 1.
}
\details{
Optional arguments

\itemize{
\item \code{new.length} the new length of the unified vector. This will be rounded upwards to a power of 2.
\item \code{force} force the calculation of large designs
\item \code{reference} the name of the variable used as reference for the fixed aspect ratio
\item \code{scale} a named vector with the scales used of the fixed aspect ratio. \code{reference = "X"} and \code{scale = c(Y = 2)} will fixed aspect ratio so that \code{Y / X = 2}
}

\itemize{
\item \code{cellsize} the cellsize for the spatial grid. Defaults to one thirth of the minimum distance between nearest neighbours
}
}
\keyword{methods}
