% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NdGRTS.R, R/NdGRTS-list.R,
%   R/NdGRTS-data_frame.R
\docType{methods}
\name{NdGRTS-methods}
\alias{NdGRTS-methods}
\alias{NdGRTS}
\alias{NdGRTS-methods}
\alias{NdGRTS,list-method}
\alias{NdGRTS-methods}
\alias{NdGRTS,data.frame-method}
\title{Calculate the N dimensional GRTS ranking for the object}
\usage{
NdGRTS(object, ...)

\S4method{NdGRTS}{list}(object, ...)

\S4method{NdGRTS}{data.frame}(object, ...)
}
\arguments{
\item{object}{The object on which to generate the N dimensional GRTS ranking}

\item{...}{further arguments to the functions}
}
\value{
a data.frame with the design and ranking
}
\description{
While GRTS() works only on 2 dimensional data, NdGRTS() can work with several dimensions, starting from 1.
}
\keyword{methods}
