% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NdRanking.R
\name{NdRanking}
\alias{NdRanking}
\title{Workhorse for multidimensional GRTS}
\usage{
NdRanking(design, level = log2(max(design) + 1) - 1, n_var = ncol(design))
}
\arguments{
\item{design}{the design matrix: one column for each dimension. Each column holds integer values between \eqn{0} and \eqn{2^{level+1}}}

\item{level}{the number of required recursive calls}

\item{n_var}{the number of columns in the design matrix}
}
\description{
Workhorse for multidimensional GRTS
}
