% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_length.R, R/unify_length-integer.R,
%   R/unify_length-numeric.R
\docType{methods}
\name{unify_length-methods}
\alias{unify_length-methods}
\alias{unify_length}
\alias{unify_length-methods}
\alias{unify_length,integer-method}
\alias{unify_length-methods}
\alias{unify_length,integer-method}
\title{Expand a vector to given length}
\usage{
unify_length(x, new.length)

\S4method{unify_length}{integer}(x, new.length)

\S4method{unify_length}{numeric}(x, new.length)
}
\arguments{
\item{x}{the vector to expand}

\item{new.length}{the required length of the vector}
}
\value{
a data.frame with the design and ranking
}
\description{
Dummy intermediate values will be added.
}
\keyword{methods}
