\name{vanderCorput}
\alias{vanderCorput}
\title{calculates a van der Corput sequence}
\usage{
  vanderCorput(size, base = 2, start)
}
\arguments{
  \item{size}{the length of the resulting sequence}

  \item{base}{a prime number that is used a the base}

  \item{start}{the starting position. A random value is
  used when missing.}
}
\description{
  calculates a van der Corput sequence
}

