\name{GRTS.grid}
\alias{GRTS.grid}
\title{GRTS sampling on grids}
\usage{
  GRTS.grid(spGrid, ...)
}
\arguments{
  \item{spGrid}{A \code{\link{SpatialGrid}} or
  \code{\link{SpatialGridDataFrame}} object}

  \item{...}{Ignored}
}
\value{
  a SpatialGridDataFrame object with a variable 'Ranking'
  indicating the GRTS order
}
\description{
  GRTS sampling on grids
}

